Implement setBusinessDayChecker() method
